refactor: WinningCondition 클래스 별도의 파일로 추출, 빌더패턴 적용
